lambdas/nodes/image_proxy/requirements.txt
pillow==11.0.0
aws-lambda-powertools